Move onion skin preview into separate component
