+++ src/renderer/components/animator/Preview/PreviewFrame/PreviewFrame.tsx
@@ -1,4 +1,3 @@
-import classNames from "classnames";
 import { useEffect, useRef, useState } from "react";
 import "./PreviewFrame.css";
